pyproject.toml
[project]
name = "logger"
version = "0.1.0"
description = "A personal life logging tool with AI-powered analysis using Gemini API"
readme = "README.md"
authors = [
    { name = "Better", email = "better@moego.pet" }
]
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.3.0",
    "sqlalchemy>=2.0.0",
    "click>=8.1.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "jinja2>=3.1.0",
    "pillow>=10.0.0",
    "pyperclip>=1.8.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.scripts]
logger = "logger.cli:cli"

[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"
